fixing chat axios and seting background
