+++ serafin/.gitignore
@@ -1,6 +1,8 @@
 # OSX
 #
 .DS_Store
+# .gitignore
+config.js
 
 # Xcode
 #
