+++ serafin/src/navigation/Drawer.tsx
@@ -6,8 +6,8 @@ import Boleto from '../screens/BoletoScreen';
 import Login from '../screens/Login';
 import CustomDrawerContent from './CustomDrawer';
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import NewChat from '../screens/NewChat';
-import Chat from '../screens/Chat';
+import NewChat from '../screens/chat/Chat';
+import Chat from '../screens/chat/Chat';
 import HomeNew from '../screens/home/NewHome';
 
 
@@ -62,7 +62,7 @@ export default function MyDrawer() {
       }}/>
       <Drawer.Screen 
         name="ChatBot" 
-        component={ChatApp} 
+        component={Chat} 
         options={{
           title: "ChatBot",
           drawerIcon: ({focused, size}) => (
